+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # @naparajith/tailwind-helpers
 
+## 0.2.0
+
+### Minor Changes
+
+- Added full code documentation for editor code lens
+- Generated tests for variants and concat package
+- Improve in overall performance by 1%
+
 ## 0.1.3
 
 ### Patch Changes